chore!: use u64 for shortint metadata
This switches from usize to u64 for shortint's metdata:
* Degree
* MaxDegree
* CarryModulus
* MessageModulus

The reasoning is that usize should be preferred when the value is used
as some kind of index, memory access, etc, and not numbers like these
metadata are.

This is a breaking API change
This is also a somewhat breaking serialization change
depending on the serialization format (bincode should be ok as it
encodes usize as u64)

+++ tfhe/src/integer/ciphertext/integer_ciphertext.rs
@@ -4,10 +4,7 @@ use crate::shortint::Ciphertext;
 pub trait IntegerCiphertext: Clone {
     fn blocks(&self) -> &[Ciphertext];
     fn moduli(&self) -> Vec<u64> {
-        self.blocks()
-            .iter()
-            .map(|x| x.message_modulus.0 as u64)
-            .collect()
+        self.blocks().iter().map(|x| x.message_modulus.0).collect()
     }
 
     fn from_blocks(blocks: Vec<Ciphertext>) -> Self;
@@ -87,7 +84,7 @@ impl IntegerCiphertext for CrtCiphertext {
     }
 
     fn from_blocks(blocks: Vec<Ciphertext>) -> Self {
-        let moduli = blocks.iter().map(|x| x.message_modulus.0 as u64).collect();
+        let moduli = blocks.iter().map(|x| x.message_modulus.0).collect();
         Self { blocks, moduli }
     }
